DatabaseScheme

Description

DatabaseScheme exports basic types and function for creating, storing and analyzing the DatabaseScheme for analysed queries.

Documentation

data DatabaseScheme

Type for storing the database scheme. It maps table names to column names and column names to their types.

DataType

data DataType

DataType contains all the implemented data types

Functions

databaseSchemeFromAst

databaseSchemeFromAst :: [Statement] -> DatabaseScheme

databaseSchemeFromAst creates the DatabaseScheme value based on the abstract syntax tree of the DDL.

getColumnNames

getColumnNames

Arguments

:: String
table name

-> DatabaseScheme
database scheme

-> [String]
column names

Function returns the list of column names for the given table. If the table does not exist and empty list is returned.

getTablesWithArity

getTablesWithArity :: DatabaseScheme -> [(String, Int)]

addTables

addTables :: [(String, [String])] -> DatabaseScheme -> DatabaseScheme