Translate

Description

Module exports function translateStatements, which creates a query description in the TPTP syntax.

Currently only first-order logic statements are supported and only a small subset of SQL queries.

BNF of the accepted SQL queries

Documentation

translateStatements

Arguments

:: [Statement]
List of SQL statements -- either queries or database schema description

-> Either String String
Either error or database scheme description and translated query

The translateStatements function takes two lists of statements (the abstract syntax tree of a DDL and a query) and returns either an error message, when an eror occured during the translation of the query or a TPTP translation of the queries.

Statement structure described here: https://github.com/JakeWheat/simple-sql-parser/blob/master/Language/SQL/SimpleSQL/Syntax.lhs