DatabaseScheme

Description

DatabaseScheme exports basic types and function for creating, storing and analyzing the DatabaseScheme for analysed queries.

Types

DatabaseScheme

type DatabaseScheme = Map String (Map String DataType)

Type for storing the database scheme. It maps table names to column names and column names to their types.

DataType

data DataType

DataType contains all the implemented data types

Functions

databaseSchemeFromAst

databaseSchemeFromAst :: [Statement] -> DatabaseScheme

databaseSchemeFromAst creates the DatabaseScheme value based on the abstract syntax tree of the DDL.