Translate

Description

Here is a longer description of this module, containing some commentary with some markup.

Functions

translateStatements

translateStatements

Arguments

:: [Statement]
databaseSchemeAst

-> [Statement]
queryAst

-> IO (Either String (String, String))

The translateStatements function takes two lists of statements (the abstract syntax tree of a DDL and a query) and returns either an error message, when an eror occured during the translation of the query or two strings one of which contains the description of the database scheme, and the other contains the query translated to the TPTP syntax

Statement structure described here: https://github.com/JakeWheat/simple-sql-parser/blob/master/Language/SQL/SimpleSQL/Syntax.lhs